EA: manifest dark
