+++ src/components/v2/appDetails/k8Resource/nodeType/Node.component.tsx
@@ -12,6 +12,7 @@ import { getNodeDetailTabs } from '../nodeDetail/nodeDetail.util';
 import NodeDeleteComponent from './NodeDelete.component';
 import AppDetailsStore from '../../appDetails.store';
 import { toast } from 'react-toastify';
+import { getNodeStatus } from './nodeType.util';
 
 function NodeComponent() {
     const { path, url } = useRouteMatch();
@@ -122,23 +123,6 @@ function NodeComponent() {
         }
     };
 
-    const getNodeStatus = (node: iNode) => {
-        if (node.info && node.info.length > 0) {
-            const statusReason = node.info.filter((_info) => {
-                return _info.name === 'Status Reason';
-            });
-            if (statusReason && statusReason.length > 0) {
-                return statusReason[0].value;
-            }
-        }
-        if (node.status) {
-            return node.status;
-        }
-        if (node.health?.status) {
-            return node.health?.status;
-        }
-        return '';
-    };
 
     const makeNodeTree = (nodes: Array<iNode>, showHeader?: boolean) => {
         return nodes.map((node, index) => {
